+++ parity-match/src/main/java/org/jvirtanen/parity/match/Market.java
@@ -43,7 +43,7 @@ public class Market {
      * @param side the side
      * @param size the size
      */
-    public void enter(long orderId, Side side, int size) {
+    public void enter(long orderId, Side side, long size) {
         if (orders.containsKey(orderId))
             return;
 
@@ -57,8 +57,8 @@ public class Market {
         }
     }
 
-    private void match(long orderId, Orders orders, int size) {
-        int remainingQuantity = size;
+    private void match(long orderId, Orders orders, long size) {
+        long remainingQuantity = size;
 
         Level top = orders.getBestLevel();
 
@@ -89,7 +89,7 @@ public class Market {
      * @param price the limit price
      * @param size the size
      */
-    public void enter(long orderId, Side side, long price, int size) {
+    public void enter(long orderId, Side side, long price, long size) {
         if (orders.containsKey(orderId))
             return;
 
@@ -103,8 +103,8 @@ public class Market {
         }
     }
 
-    private void buy(long orderId, long price, int size) {
-        int remainingQuantity = size;
+    private void buy(long orderId, long price, long size) {
+        long remainingQuantity = size;
 
         Level top = asks.getBestLevel();
 
@@ -121,8 +121,8 @@ public class Market {
         }
     }
 
-    private void sell(long orderId, long price, int size) {
-        int remainingQuantity = size;
+    private void sell(long orderId, long price, long size) {
+        long remainingQuantity = size;
 
         Level top = bids.getBestLevel();
 
@@ -151,12 +151,12 @@ public class Market {
      * @param orderId the order identifier
      * @param size the new size
      */
-    public void cancel(long orderId, int size) {
+    public void cancel(long orderId, long size) {
         Order order = orders.get(orderId);
         if (order == null)
             return;
 
-        int remainingQuantity = order.getRemainingQuantity();
+        long remainingQuantity = order.getRemainingQuantity();
 
         if (size >= remainingQuantity)
             return;
